Projects/Advanced_text_to_sql/requirements.txt
streamlit
python-dotenv
openai
langchain_community
langchain_core
langchain_openai
cohere
pandas
typing-extensions
chromadb
